api/pyproject.toml
[tool.poetry]
name = "picsearch"
version = "0.1.0"
description = ""
authors = ["mammrezesk <mohammadreza.eskandari@divar.ir>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
torch = "^2.4.1"
torchvision = "^0.19.1"
ftfy = "^6.2.3"
regex = "^2024.9.11"
tqdm = "^4.66.5"
clip = { git = "https://github.com/openai/CLIP.git" }
pinecone-client = "^5.0.1"
requests = "^2.32.3"
django-environ = "^0.11.2"
apache-airflow = "^2.10.2"
mongoengine = "^0.29.1"
boto3 = "^1.35.37"
meilisearch = "^0.31.5"
openai = "^1.51.2"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
pymongo = "^4.10.1"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
